Finshing up with authorization

+++ Application/Features/Restaurants/Commands/Delete/DeleteRestaurantCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using Domain.Entities;
+using Domain.Enums;
 using Domain.Exceptions;
 using Domain.Interfaces;
 using MediatR;
@@ -6,7 +7,10 @@ using Microsoft.Extensions.Logging;
 
 namespace Application.Features.Restaurants.Commands.Delete
 {
-    public class DeleteRestaurantCommandHandler(ILogger<DeleteRestaurantCommand> logger, IRestaurantsRepository restaurantsRepository) : IRequestHandler<DeleteRestaurantCommand, Unit>
+    public class DeleteRestaurantCommandHandler(
+        ILogger<DeleteRestaurantCommand> logger,
+        IRestaurantsRepository restaurantsRepository,
+        IRestaurantAuthorizationServices restaurantAuthorizationServices) : IRequestHandler<DeleteRestaurantCommand, Unit>
     {
         public async Task<Unit> Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
         {
@@ -17,6 +21,12 @@ namespace Application.Features.Restaurants.Commands.Delete
             {
                 logger.LogWarning("Restaurant with ID: {Id} not found", request.Id);
                 throw new NotFoundException(nameof(Restaurant), request.Id.ToString());
+            }
+
+            if (!restaurantAuthorizationServices.Authorize(restaurant, ResourceOperationEnum.Delete))
+            {
+                logger.LogWarning("Unauthorized attempt to delete restaurant with ID: {Id}", request.Id);
+                throw new AccessForbiddenException();
             }
             await restaurantsRepository.DeleteRestaurantAsync(restaurant);
             return Unit.Value;
